requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
beautifulsoup4==4.12.3
lxml==5.3.0
regex==2024.11.6
selenium==4.26.1
webdriver-manager==4.0.2
tensorflow==2.18.0
Pillow==11.0.0
streamlit==1.40.1
transformers==4.46.3
torch==2.5.1
datasets==3.1.0
kaggle==1.6.17
python-dotenv==1.0.1
